feat: use half-duplex UART to support bidirectional TMC2209 comms
an external pull-up resistor on the UART bus is now required.

reply microsecond delays need to be tuned as they are very
long to ensure we don't miss anything.

echo handling needed to be removed as the STM32H5 half-duplex
implementation does not need it and rather suffers from it.